refactor: 공통 변수 동일 dto로 이전
