+++ src/main/java/leets/weeth/domain/schedule/application/dto/MeetingDTO.java
@@ -1,28 +1,8 @@
 package leets.weeth.domain.schedule.application.dto;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import org.springframework.format.annotation.DateTimeFormat;
-
 import java.time.LocalDateTime;
 
 public class MeetingDTO {
-
-    public record ResponseAll(
-            Long id,
-            String title,
-            String content,
-            String location,
-            LocalDateTime start,
-            LocalDateTime end,
-            Integer weekNumber,
-            Integer cardinal,
-            Integer code,
-            String name,
-            Integer memberCount,
-            LocalDateTime createdAt,
-            LocalDateTime modifiedAt
-    ) {}
 
     public record Response(
             Long id,
@@ -32,29 +12,9 @@ public class MeetingDTO {
             LocalDateTime start,
             LocalDateTime end,
             String name,
-            Integer memberCount,
             String requiredItem,
             LocalDateTime createdAt,
             LocalDateTime modifiedAt
     ) {}
 
-    public record Save(
-            @NotBlank String title,
-            @NotBlank String content,
-            @NotBlank String location,
-            @NotNull @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime start,
-            @NotNull @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime end,
-            @NotNull Integer weekNumber,
-            @NotNull Integer cardinal
-    ) {}
-
-    public record Update(
-            @NotBlank String title,
-            @NotBlank String content,
-            @NotBlank String location,
-            @NotNull @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime start,
-            @NotNull @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime end,
-            @NotNull Integer weekNumber,
-            @NotNull Integer cardinal
-    ) {}
 }
